Zmiana typów liczb

+++ src/main/java/pl/javastart/task/Point.java
@@ -1,10 +1,10 @@
 package pl.javastart.task;
 
 public class Point {
-    double x;
-    double y;
+    int x;
+    int y;
 
-    public Point(double x, double y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
